+++ src/client/components/DirectoryView/components/Breadcrumbs/Breadcrumbs.tsx
@@ -1,12 +1,12 @@
 import { House } from "@phosphor-icons/react";
-import { BreadcrumbItem } from "~/client/components/Breadcrumbs/components/BreadcrumbItem";
+import { BreadcrumbItem } from "~/client/components/DirectoryView/components/Breadcrumbs/components/BreadcrumbItem";
 
 interface Props {
   path: string[];
 }
 
 export const Breadcrumbs = ({ path }: Props) => (
-  <nav className="flex rounded-t-md bg-base-200 p-2">
+  <nav className="flex h-8 rounded-t-md bg-base-200 px-2">
     <ol className="flex items-center gap-1">
       <BreadcrumbItem path={["/"]} title={<House weight="fill" size={16} />} />
       {path?.map((dirname, index) => (
