+++ src/client/components/DirectoryView/components/Breadcrumbs/components/BreadcrumbItem.tsx
@@ -1,7 +1,7 @@
 import { type ReactNode } from "react";
 import { CaretRight } from "@phosphor-icons/react";
-import { BreadcrumbLink } from "~/client/components/Breadcrumbs/components/BreadcrumbLink";
-import { BreadcrumbText } from "~/client/components/Breadcrumbs/components/BreadcrumbText";
+import { BreadcrumbLink } from "~/client/components/DirectoryView/components/Breadcrumbs/components/BreadcrumbLink";
+import { BreadcrumbText } from "~/client/components/DirectoryView/components/Breadcrumbs/components/BreadcrumbText";
 
 interface Props {
   title: ReactNode;
