+++ src/client/components/DirectoryView/components/Breadcrumbs/components/BreadcrumbText.tsx
